Backend/requirements_production.txt
asgiref==3.11.0
dj-database-url==3.1.1
Django==6.0.1
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.5
joblib==1.5.3
numpy==2.4.0
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.3
redis==7.1.0
hiredis==3.3.0
whitenoise==6.8.0
